Add files via upload
This is a file to implement the communication with Ditto service deployed on private cloud and listen for the SSE service for updates on the instances. Each update is stored into the mySQL DB to allow the Dashboard to visualize real-time updates on the RFID sensors data and position of the UWB sensors.